% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gxe.R
\name{plot_reaction_norms}
\alias{plot_reaction_norms}
\title{Plot Reaction Norms (GxE)}
\usage{
plot_reaction_norms(object, x, group, response, add_points = TRUE, ...)
}
\arguments{
\item{object}{A semx_fit object.}

\item{x}{The name of the environmental variable (x-axis).}

\item{group}{The name of the grouping variable (genotype/individual).}

\item{response}{The name of the response variable.}

\item{add_points}{Logical. If TRUE, adds observed data points.}

\item{...}{Additional arguments passed to plot.}
}
\description{
Visualizes the interaction between genotype and environment by plotting
reaction norms (phenotype vs environment) for each genotype.
}
