% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossed.R
\name{print.semx_crossed_fit}
\alias{print.semx_crossed_fit}
\title{Print method for semx_crossed_fit}
\usage{
\method{print}{semx_crossed_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{semx_crossed_fit}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for semx_crossed_fit
}
