% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semx.R
\name{semx_covariance_weights}
\alias{semx_covariance_weights}
\title{Extract covariance weights}
\usage{
semx_covariance_weights(fit, cov_id)
}
\arguments{
\item{fit}{A semx_fit object.}

\item{cov_id}{Covariance identifier.}
}
\value{
A list with sigma_sq and weights, or NULL.
}
\description{
Extract covariance weights
}
