% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semx.R
\name{semx_cv_genomic_prediction}
\alias{semx_cv_genomic_prediction}
\title{Cross-validation for genomic prediction}
\usage{
semx_cv_genomic_prediction(
  model,
  data,
  outcome,
  folds = 5,
  seed = 42,
  options = NULL,
  optimizer_name = "lbfgs"
)
}
\arguments{
\item{model}{A semx_model object.}

\item{data}{A data.frame.}

\item{outcome}{Name of the outcome variable.}

\item{folds}{Number of folds.}

\item{seed}{Random seed.}

\item{options}{Optimization options.}

\item{optimizer_name}{Optimizer name.}
}
\value{
List of metrics.
}
\description{
Cross-validation for genomic prediction
}
