% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semx.R
\name{semx_diagnostics}
\alias{semx_diagnostics}
\title{Compute model diagnostics}
\usage{
semx_diagnostics(fit, model, data)
}
\arguments{
\item{fit}{A FitResult object.}

\item{model}{A semx_model object.}

\item{data}{A list or data.frame containing the data.}
}
\value{
A list containing model diagnostics (implied moments, residuals, SRMR).
}
\description{
Compute model diagnostics
}
