% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semx.R
\name{semx_extract_heritability}
\alias{semx_extract_heritability}
\title{Extract heritability}
\usage{
semx_extract_heritability(fit, genetic_component, residual_component)
}
\arguments{
\item{fit}{A semx_fit object.}

\item{genetic_component}{Name of the genetic variance parameter.}

\item{residual_component}{Name of the residual variance parameter.}
}
\value{
Heritability estimate.
}
\description{
Extract heritability
}
