% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semx.R
\name{semx_fit}
\alias{semx_fit}
\title{Fit a semx model}
\usage{
semx_fit(
  model,
  data,
  options = NULL,
  optimizer_name = "lbfgs",
  fixed_covariance_data = NULL,
  estimation_method = "ML"
)
}
\arguments{
\item{model}{A semx_model object.}

\item{data}{A list or data.frame containing the data.}

\item{options}{Optimization options.}

\item{optimizer_name}{Name of the optimizer ("lbfgs" or "gd").}

\item{fixed_covariance_data}{Optional list of fixed covariance matrices.}

\item{estimation_method}{Estimation method ("ML" or "REML").}
}
\value{
A FitResult object.
}
\description{
Fit a semx model
}
