% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semx.R
\name{semx_modification_indices}
\alias{semx_modification_indices}
\title{Compute modification indices}
\usage{
semx_modification_indices(fit, model = NULL, data = NULL)
}
\arguments{
\item{fit}{A FitResult object or semx_fit object.}

\item{model}{A semx_model object (optional if fit is semx_fit).}

\item{data}{A list or data.frame containing the data (optional if fit is semx_fit).}
}
\value{
A data.frame containing modification indices.
}
\description{
Compute modification indices
}
