% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path.R
\name{semx_plot_path}
\alias{semx_plot_path}
\title{Plot Path Diagram}
\usage{
semx_plot_path(fit, ...)
}
\arguments{
\item{fit}{A semx_fit object.}

\item{...}{Additional arguments passed to DiagrammeR::grViz.}
}
\value{
A DiagrammeR object.
}
\description{
Plot Path Diagram
}
\keyword{internal}
