% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{semx_predict_cif}
\alias{semx_predict_cif}
\title{Predict Cumulative Incidence Function}
\usage{
semx_predict_cif(fit, newdata, times, event_outcome, competing_outcomes)
}
\arguments{
\item{fit}{A semx_fit object.}

\item{newdata}{A data.frame.}

\item{times}{Numeric vector of times.}

\item{event_outcome}{Name of the event outcome.}

\item{competing_outcomes}{Vector of competing outcome names.}
}
\value{
A data.frame of CIF values.
}
\description{
Predict Cumulative Incidence Function
}
