% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{semx_predict_survival}
\alias{semx_predict_survival}
\title{Predict survival probabilities}
\usage{
semx_predict_survival(fit, newdata, times, outcome)
}
\arguments{
\item{fit}{A semx_fit object.}

\item{newdata}{A data.frame.}

\item{times}{Numeric vector of times.}

\item{outcome}{Name of the outcome variable.}
}
\value{
A data.frame of survival probabilities.
}
\description{
Predict survival probabilities
}
