% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{semx_ranef}
\alias{semx_ranef}
\title{Extract Random Effects}
\usage{
semx_ranef(object, ...)
}
\arguments{
\item{object}{A semx_fit object}

\item{...}{Additional arguments (ignored)}
}
\value{
A list of matrices, one per random effect block.
}
\description{
Extract Random Effects
}
