% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semx.R
\name{semx_variance_components}
\alias{semx_variance_components}
\title{Extract variance components}
\usage{
semx_variance_components(fit)
}
\arguments{
\item{fit}{A semx_fit object.}
}
\value{
A data.frame with variance components.
}
\description{
Extract variance components
}
