% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossed.R
\name{summary.semx_crossed_fit}
\alias{summary.semx_crossed_fit}
\title{Summary method for semx_crossed_fit}
\usage{
\method{summary}{semx_crossed_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{semx_crossed_fit}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for semx_crossed_fit
}
