% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semx.R
\name{semx_modification_indices}
\alias{semx_modification_indices}
\title{Compute modification indices}
\usage{
semx_modification_indices(fit, model, data)
}
\arguments{
\item{fit}{A FitResult object.}

\item{model}{A semx_model object.}

\item{data}{A list or data.frame containing the data.}
}
\value{
A data.frame containing modification indices.
}
\description{
Compute modification indices
}
