% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semx.R
\name{semx_standardized_solution}
\alias{semx_standardized_solution}
\title{Compute standardized estimates}
\usage{
semx_standardized_solution(fit, model, data)
}
\arguments{
\item{fit}{A FitResult object.}

\item{model}{A semx_model object.}

\item{data}{A list or data.frame containing the data.}
}
\value{
A list containing standardized estimates.
}
\description{
Compute standardized estimates
}
